added a play_again function
As you can see it's a bit weird as I had to write it that way because if I do input(spec_print(...)) the print will have be like

"Do you want to play again? (y/n): None" 

So I had to print the "Do you want to play again? (y/n): " first then initialize a variable that will take an input, the new_list is there to make the illusion that it's in one function.